Create a non-ui ScriptEditor file for interacting with the script editor when the UI component is not being displayed
