+++ src/ScriptEditor/ui/ScriptEditorRoot.tsx
@@ -12,19 +12,15 @@ import { Page } from "../../ui/Router";
 import { dialogBoxCreate } from "../../ui/React/DialogBox";
 import { checkInfiniteLoop } from "../../Script/RamCalculations";
 
-import { ns, enums } from "../../NetscriptFunctions";
 import { Settings } from "../../Settings/Settings";
 import { iTutorialNextStep, ITutorial, iTutorialSteps } from "../../InteractiveTutorial";
 import { debounce } from "lodash";
 import { saveObject } from "../../SaveObject";
-import { loadThemes, makeTheme, sanitizeTheme } from "./themes";
 import { GetServer } from "../../Server/AllServers";
 
 import { PromptEvent } from "../../ui/React/PromptManager";
 
-import libSource from "!!raw-loader!../NetscriptDefinitions.d.ts";
 import { useRerender } from "../../ui/React/hooks";
-import { NetscriptExtra } from "../../NetscriptFunctions/Extra";
 
 import { dirty, getServerCode } from "./utils";
 import { OpenScript } from "./OpenScript";
@@ -40,24 +36,6 @@ interface IProps {
   hostname: string;
   vim: boolean;
 }
-
-// TODO: try to remove global symbols
-let symbolsLoaded = false;
-const apiKeys: string[] = [];
-export function SetupTextEditor(): void {
-  // Function for populating apiKeys using a given layer of the API.
-  const api = { args: [], pid: 1, enums, ...ns };
-  const hiddenAPI = NetscriptExtra();
-  function populate(apiLayer: object = api) {
-    for (const [apiKey, apiValue] of Object.entries(apiLayer)) {
-      if (apiLayer === api && apiKey in hiddenAPI) continue;
-      apiKeys.push(apiKey);
-      if (typeof apiValue === "object") populate(apiValue);
-    }
-  }
-  populate();
-}
-
 const openScripts: OpenScript[] = [];
 let currentScript: OpenScript | null = null;
 
@@ -136,47 +114,6 @@ function Root(props: IProps): React.ReactElement {
     startUpdatingRAM();
     debouncedCodeParsing(newCode);
   };
-
-  // How to load function definition in monaco
-  // https://github.com/Microsoft/monaco-editor/issues/1415
-  // https://microsoft.github.io/monaco-editor/api/modules/monaco.languages.html
-  // https://www.npmjs.com/package/@monaco-editor/react#development-playground
-  // https://microsoft.github.io/monaco-editor/playground.html#extending-language-services-custom-languages
-  // https://github.com/threehams/typescript-error-guide/blob/master/stories/components/Editor.tsx#L11-L39
-  // https://blog.checklyhq.com/customizing-monaco/
-  // Before the editor is mounted
-  function beforeMount(): void {
-    if (symbolsLoaded) return;
-    // Setup monaco auto completion
-    symbolsLoaded = true;
-    (async function () {
-      // We have to improve the default js language otherwise theme sucks
-      const jsLanguage = monaco.languages.getLanguages().find((l) => l.id === "javascript");
-      // Unsupported function is not exposed in monaco public API.
-      const l = await (jsLanguage as any).loader();
-      // replaced the bare tokens with regexes surrounded by \b, e.g. \b{token}\b which matches a word-break on either side
-      // this prevents the highlighter from highlighting pieces of variables that start with a reserved token name
-      l.language.tokenizer.root.unshift([new RegExp("\\bns\\b"), { token: "ns" }]);
-      for (const symbol of apiKeys)
-        l.language.tokenizer.root.unshift([new RegExp(`\\b${symbol}\\b`), { token: "netscriptfunction" }]);
-      const otherKeywords = ["let", "const", "var", "function"];
-      const otherKeyvars = ["true", "false", "null", "undefined"];
-      otherKeywords.forEach((k) =>
-        l.language.tokenizer.root.unshift([new RegExp(`\\b${k}\\b`), { token: "otherkeywords" }]),
-      );
-      otherKeyvars.forEach((k) =>
-        l.language.tokenizer.root.unshift([new RegExp(`\\b${k}\\b`), { token: "otherkeyvars" }]),
-      );
-      l.language.tokenizer.root.unshift([new RegExp("\\bthis\\b"), { token: "this" }]);
-    })();
-
-    const source = (libSource + "").replace(/export /g, "");
-    monaco.languages.typescript.javascriptDefaults.addExtraLib(source, "netscript.d.ts");
-    monaco.languages.typescript.typescriptDefaults.addExtraLib(source, "netscript.d.ts");
-    loadThemes(monaco.editor.defineTheme);
-    sanitizeTheme(Settings.EditorTheme);
-    monaco.editor.defineTheme("customTheme", makeTheme(Settings.EditorTheme));
-  }
 
   // When the editor is mounted
   function onMount(editor: IStandaloneCodeEditor): void {
@@ -459,7 +396,7 @@ function Root(props: IProps): React.ReactElement {
           onTabUpdate={onTabUpdate}
         />
         <div style={{ flex: "0 0 5px" }} />
-        <Editor beforeMount={beforeMount} onMount={onMount} onChange={updateCode} />
+        <Editor onMount={onMount} onChange={updateCode} />
 
         {VimStatus}
