+++ src/engine.tsx
@@ -9,7 +9,6 @@ import { staneksGift } from "./CotMG/Helper";
 import { processPassiveFactionRepGain, inviteToFaction } from "./Faction/FactionHelpers";
 import { Router } from "./ui/GameRoot";
 import { Page } from "./ui/Router";
-import { SetupTextEditor } from "./ScriptEditor/ui/ScriptEditorRoot";
 import "./PersonObjects/Player/PlayerObject"; // For side-effect of creating Player
 
 import {
@@ -377,7 +376,6 @@ const Engine: {
       // Start interactive tutorial
       iTutorialStart();
     }
-    SetupTextEditor();
   },
 
   start: function () {
